third_party/blink/web_tests/fast/forms/number/number-outofrange-expected.txt

Test for spinbutton behavior for out-of-range values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Pressing the down arrow key on an input field of which value is lower than the minimum:
PASS lower.value is unchanged
Pressing the up arrow key on the input:
PASS lower.value is lower.min
Pressing the up arrow key on an input field of which value is higher than the maximum:
PASS higher.value is unchanged
Pressing the down arrow key on the input:
PASS higher.value is higher.max
PASS successfullyParsed is true

TEST COMPLETE
